% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-themes-and-palettes.R
\name{acc.pal}
\alias{acc.pal}
\title{acc.pal}
\usage{
acc.pal(n = 6, names = F)
}
\description{
Pretty / jewel-toned color palette designed by accessibility-minded designers
from UK civil service data network:
https://analysisfunction.civilservice.gov.uk/.
}
