% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-themes-and-palettes.R
\name{add.map.theme}
\alias{add.map.theme}
\title{add.map.theme}
\usage{
add.map.theme(
  fontsize = 11,
  leg.pos = "right",
  lmargin = margin(4, 4, 4, 4),
  ...
)
}
\description{
Add base map theme.
}
