% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{attach.geos}
\alias{attach.geos}
\title{attach.geos}
\usage{
attach.geos(x, geoid.col = "geoid", query_fcn = tigris::tracts, crs = 4326)
}
\arguments{
\item{x}{df with at least 1 column containing state+county geoid hierarchy (so
that first 5 characters are state and county fips codes)}

\item{geoid.col}{column representing geoids}

\item{query_fcn}{function to query geos from census api. Defau9lt gets tracts}

\item{...}{passed onto \code{query_fcn}}
}
\value{
\code{x}, as \code{sf} object with geometries attached
}
\description{
attach.geos
}
