% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-helpers.R
\name{bbox2ggcrop}
\alias{bbox2ggcrop}
\title{bbox2ggcrop}
\usage{
bbox2ggcrop(sfx, crs = 4326, clip = "on")
}
\arguments{
\item{sfx}{bbox or sf object}

\item{clip}{passed onto \code{ggplot2::coord_sf}}
}
\description{
Wraps \code{coord_sf}, which can be used to crop a map made in ggplot. Uses an sf object
or bbox to set crop area.
}
