% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet helpers.R
\name{bbox_to_lnglat}
\alias{bbox_to_lnglat}
\title{bbox_to_lnglat}
\usage{
bbox_to_lnglat(bbox, padding = 1)
}
\arguments{
\item{padding}{padding to put around bbox. Units set by map crs}

\item{st_bbox}{output}
}
\description{
Leaflet fcns like \code{leaflet::flyToBounds} and \code{leaflet::setMaxBounds} require bounds in
a format slightly different than that returned from sf::st_bbox. This
helper turns output of \code{sf::st_bbox} to an input for a leaflet zoom function. Can add a
small pad around the bbox as well.
}
