% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning and formatting.R
\name{bin.var_format}
\alias{bin.var_format}
\title{bin.var_format}
\usage{
bin.var_format(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{
  Arguments passed on to \code{\link[=get_breaks]{get_breaks}}, \code{\link[=bin_from_breaks]{bin_from_breaks}}
  \describe{
    \item{\code{n_breaks}}{target number of breaks for x}
    \item{\code{highlight_top_percent}}{Whether to break out an additional bin for the top
1\\%. Helpful to break out upper outliers.}
    \item{\code{format_breaks}}{whether to create "x-y" type character labels for buckets rather
than default labels from \code{cut}}
    \item{\code{digits}}{Digits to pass on to \code{q.format} if using formatting}
  }}
}
\description{
Bins a sequential variable based on value ranges. It can often make visualizations
easier to look at by allowing more striking differences and keeping outliers from
dominating the color scale. Can return formatted values if
\code{use_labels} is TRUE. the highlight_top parameter ensures a separate bin for
top 1\% The function wraps \code{get_breaks} and \code{bin_from_breaks}, which can
also be used separately
}
