% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-themes-and-palettes.R
\name{breaks_log10}
\alias{breaks_log10}
\title{breaks_log10}
\usage{
breaks_log10(x, break.intervals = c(1))
}
\arguments{
\item{x}{vector mapped to x or y scales.}

\item{break.intervals}{A vector of numerics between \link{1,9} for there to be
major breaks at every power of 10. I.e. every 1e6, 1e7 if left at just one,
but 1e6, 3e6, 1e6, 3e7 if includes \code{c(1, 3)}.}
}
\description{
Manually adds breaks at very power of 10 for log-scale axes.
}
\details{
solution adapted from https://r-graphics.org/recipe-axes-axis-log-ticks
}
