% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning-and-formatting.R
\name{cap.at.quantile}
\alias{cap.at.quantile}
\title{cap.at.quantile}
\usage{
cap.at.quantile(x, ptile = 0.95)
}
\arguments{
\item{x}{values}

\item{ptile}{percentile at which to cap \code{x}}
}
\description{
Caps values of a vector at a given percentile. Can be helpful for visualizing
data with long-tail distributions, especially when color is mapped to the
long tail.
}
