% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping helpers.R
\name{cntr2bbx}
\alias{cntr2bbx}
\title{cntr2bbx}
\usage{
cntr2bbx(cntr, buffer, crs)
}
\arguments{
\item{cntr}{centroid}

\item{buffer}{buffer radius around centroid in km}
}
\description{
Simple wrapper that converts a centroid to a bbox. Useful for creating ggplots at
various zoom levels, or cropping to an area of interest based on a focal point or
centroid.
}
