% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping helpers.R
\name{col2opacity}
\alias{col2opacity}
\title{col2opacity}
\usage{
col2opacity(x, n_breaks = 5, opacity_range = c(0.35, 0.95), ...)
}
\arguments{
\item{x}{numeric vector to interpolate opacities for.}

\item{n_breaks}{number of breaks cut values into (before using floor)}

\item{opacity_range}{Bounds opacities between these values. Acts as floor/ceiling
as opposed to range to interpolate along.}
}
\value{
numeric vector that can be used to represent other values through opacity.
}
\description{
Designed to interpolate numerics to opacities for a map or other visualization.
}
