% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet helpers.R
\name{create_leaflet_base}
\alias{create_leaflet_base}
\title{create_leaflet_base}
\usage{
create_leaflet_base()
}
\description{
Creates base leaflet and sets options for app. Configures layers so that text
labels on map will appear above data added to base tiles, but the base tiles for
landscape and water appear below.
}
\details{
See discussion at:
https://leafletjs.com/examples/map-panes/ or
https://stackoverflow.com/questions/54667968/controlling-the-z-index-of-a-leaflet-heatmap-in-r/54676391
for addMapPanel/zIndex details
}
