% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-helpers.R
\name{dot.map.template}
\alias{dot.map.template}
\title{dot.map.template}
\usage{
dot.map.template(dots, bbx = NULL, group.col = "group", size = 1, shape = 20)
}
\arguments{
\item{dots}{an sf object long by a group column}

\item{group.col}{group column to map by}

\item{bbox}{optional bbx to crop/transform to}
}
\description{
dot.map.template
}
