% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{flexible.spatial.filter}
\alias{flexible.spatial.filter}
\title{flexible.spatial.filter}
\usage{
flexible.spatial.filter(
  x,
  polys,
  subset.approach = c("intersects", "crop"),
  ...
)
}
\arguments{
\item{subset.approach}{spatial filter approach; one of "intersects" or "crop".
Intersection or cropping to bbox.}
}
\description{
Helper fcn that guesses appropriate spatial filter for subsetting polygons that
overlap with another sf object \code{x}, based on x's geometry type. If \code{x} contains
points, other polygon layer is subsetted with \code{st_intersects}; if \code{x} is polygons,
the other polygon is converted to points with \code{st_points_on_surface} first.
}
\details{
Currently doesn't work with non-coterminous polys, but may add that later.
}
