% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other shpfile queries.R
\name{get.NHPN}
\alias{get.NHPN}
\title{These functions use non-tigris queries that might be more specialized, but I like
for my personal convenience.
get.NHPN}
\usage{
get.NHPN(
  sfx = NULL,
  dropbox.dir = Sys.getenv("drop_dir"),
  subdir = "shapefiles/nhpn/"
)
}
\arguments{
\item{sfx}{if not NULL, an \code{sf} object to use to crop hwy data}

\item{dropbox.dir, subdir}{concatenated to get to nhpn data}
}
\description{
Get National Highway Planning Network highway data.
}
