% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning-and-formatting.R
\name{get.quantile.breaks}
\alias{get.quantile.breaks}
\title{get.quantile.breaks}
\usage{
get.quantile.breaks(x, n_breaks = 6, highlight_top_percent = F, ...)
}
\arguments{
\item{x}{numeric vector}

\item{n_breaks}{target number of breaks for x}

\item{highlight_top_percent}{Whether to break out an additional bin for the top
1\%. Helpful to break out upper outliers.}
}
\description{
bins a sequential variable based on value ranges. It can often
make visualizations easier to look at by allowing more striking differences
and keeping extremes from dominating the color scale. the highlight_top
parameter ensures a separate bin for top 1\%. Called from
quantile.bin_format.
}
