% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times-series-plotting.R
\name{get.recessions.start.ends}
\alias{get.recessions.start.ends}
\title{get.recessions.start.ends}
\usage{
get.recessions.start.ends(recessions = NULL)
}
\arguments{
\item{recessions}{as pulled from \code{fredr::fredr(series_id = 'JHDUSRGDPBR')},
with additional arguments for month/qtr/annual or observation start.}
}
\description{
Gets recession starts and endpoints; queries from FRED series "JHDUSRGDPBR"
and sets up a tibble w start and end dates as two columns.
}
