% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-helpers.R
\name{get.stamen.bkg}
\alias{get.stamen.bkg}
\title{get.stamen.bkg}
\usage{
get.stamen.bkg(sfx, maptype = "toner-background", zoom = 10, ...)
}
\description{
Gets stamen base tiles based on supplied bbox of \code{sf} object.
}
\details{
Unfortunately, ggmap has some idiosyncracies with geom_sf. If combining these into
single plot, remember:
\enumerate{
\item include \code{inherit.aes = F} in geom_sf layer
\item Match sf crs to that of stamen tiles. Should be epsg = 4326 (but looks like
google maps uses different)
}
}
