% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning and formatting.R
\name{get_breaks}
\alias{get_breaks}
\title{get_breaks}
\usage{
get_breaks(x, n_breaks = 6, highlight_top_percent = F, ...)
}
\arguments{
\item{x}{numeric vector}

\item{n_breaks}{target number of breaks for x}

\item{highlight_top_percent}{Whether to break out an additional bin for the top
1\\%. Helpful to break out upper outliers.}
}
\description{
bins a seqential variable based on value ranges. It can often
make visualisations easier to look at by allowing more striking differences
and keeping extremes from dominating the color scale. the highlight_top
parameter ensures a separate bin for top 1\%. This is called from
bin.var_format.
}
