% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning and formatting.R
\name{get_mean_from_interval}
\alias{get_mean_from_interval}
\title{get_mean_from_interval}
\usage{
get_mean_from_interval(interval)
}
\arguments{
\item{interval}{interval in form (xx,yy) or (xx,yy]}
}
\description{

}
