% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times-series-plotting.R
\name{gg.recessions.ts}
\alias{gg.recessions.ts}
\title{gg.recessions.ts}
\usage{
gg.recessions.ts(recessions = NULL, rcolor = "grey90")
}
\arguments{
\item{recessions}{A time series of recessions (or other periodizations) to
add to ggplot. Expects a \code{date} column.}

\item{rcolor}{color for bars indicating recessions etc.}

\item{...}{passed onto \code{fredr::fredr}; months or quarters and obs start may
be helpful.}
}
\description{
Adds colored bars to a (time-series) ggplot, as to indicate recessions. By
default queries recessions from FRED series "JHDUSRGDPBR."
}
