% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning-and-formatting.R
\name{ggcapped.labels}
\alias{ggcapped.labels}
\title{ggcapped.labels}
\usage{
ggcapped.labels(x, tails = "one")
}
\description{
Use in a ggplot2 scales for capped labels. Intended to be passed onto
\code{labels} argument. I.e., scale_fill_viridis_c(labels = visaux::ggcapped.labels(x))
}
