% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping helpers.R
\name{ggsave.hirez}
\alias{ggsave.hirez}
\title{ggsave.hirez}
\usage{
ggsave.hirez(
  plot,
  dir,
  fn,
  ext = "png",
  height = 7.5,
  units = "in",
  dpi = 340,
  ...
)
}
\arguments{
\item{dir, fn}{directory and filename to save to}
}
\description{
Wraps ggsave with some defaults I'm finding sensible.
}
