% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet helpers.R
\name{iterative_leaflet_draw}
\alias{iterative_leaflet_draw}
\title{i think these two should go away. Learning process!
iterative_leaflet_draw}
\usage{
iterative_leaflet_draw(leaflet.proxy, st_df, tooltips, ..., bkdwn_size = 50)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=leaflet_draw]{leaflet_draw}}
  \describe{
    \item{\code{leaflet.proxy}}{Leaflet proxy object}
    \item{\code{st_df}}{sf object to map}
    \item{\code{tooltips}}{A list of strings with html}
    \item{\code{pal}}{A function to create color palette from; i.e., a all to
leaflet::colorFactor.}
    \item{\code{opacity}}{Opacity from 0-100}
    \item{\code{var.name}}{Title of column to color map based on, possibly created with
bin.var_format or bin_from_breaks.}
  }}

\item{bkdwn_size}{Number of rows of sf object to draw at a time.}
}
\description{
clear data, get number of iterations, draw first one. Return 0 if there's
only one, otherwise loop through the rest. Idea is that it can smooth out a
render to make loading feel nicer. Passes tooltips and other graphic parameters
to \code{leaflet_draw}.
}
