% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-themes-and-palettes.R
\name{jewel.pal}
\alias{jewel.pal}
\title{jewel.pal}
\usage{
jewel.pal()
}
\description{
Returns some jewel-toney colors that can be nice for discrete data
visualization. Colorblindness simulator seems to show decent accessibility.
}
