% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet helpers.R
\name{leaflet_draw}
\alias{leaflet_draw}
\title{leaflet_draw}
\usage{
leaflet_draw(
  leaflet.proxy,
  st_df,
  tooltips,
  pal,
  opacity = 60,
  var.name = "binned_x",
  outlines = NULL,
  width = 0.5,
  popup_btns = NULL
)
}
\arguments{
\item{leaflet.proxy}{Leaflet proxy object}

\item{st_df}{sf object to map}

\item{tooltips}{A list of strings with html}

\item{pal}{A function to create color palette from; i.e., a all to
leaflet::colorFactor.}

\item{opacity}{Opacity from 0-100}

\item{var.name}{Title of column to color map based on, possibly created with
bin.var_format or bin_from_breaks.}
}
\description{
Wraps leaflet call to create a choropleth from sf object.
}
