% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning-and-formatting.R
\name{linebreaks2string}
\alias{linebreaks2string}
\title{linebreaks2string}
\usage{
linebreaks2string(x, line.char.len = 80, clean.ws = T, line.break = "\\n")
}
\description{
Removes white space (if \code{clean.ws} is true), then adds linebreaks at uniform
intervals. Can deal with very long ggplot captions, for example
}
