% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-helpers.R
\name{n.categories.discrete.palette}
\alias{n.categories.discrete.palette}
\title{n.categories.discrete.palette}
\usage{
n.categories.discrete.palette(
  ncats,
  base.pal = RColorBrewer::brewer.pal(6, "Dark2")
)
}
\description{
It's possible to run out of colors when visualizing with a large number of
discrete categories. This wraps the steps to interpolote a palette across
your number of categories, so variations of the colors are re-used. Helpful
}
\details{
It returns a ggplot2 layer, by default for a \code{fill} aesthetic
}
\examples{
x <- factor(rnorm(30))
scale_fill_manual(values = n.categories.discrete.palette(length(x)))

}
