% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning-and-formatting.R
\name{number.to.formatted.string}
\alias{number.to.formatted.string}
\title{number.to.formatted.string}
\usage{
number.to.formatted.string(
  x,
  format.fcn = scales::label_comma(accuracy = 0.1, drop0trailing = T),
  pluralize = F
)
}
\arguments{
\item{x}{A number to format (not a vector)}

\item{format.fcn}{A formatting function to apply to number part of output.}

\item{pluralize}{Whether to pluralize; i.e., "million(s)"}
}
\description{
Divides by the largest of thousand, million, billion, trillion and appends
string to modify units.
}
