% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{parks.wrapper}
\alias{parks.wrapper}
\title{parks.wrapper}
\usage{
parks.wrapper(x = NULL, statefps = NULL, ...)
}
\arguments{
\item{x}{object to derive overlapping statefps from, if statefps is left as null. Aiddtional}

\item{statefps}{statefps to get parks for}
}
\description{
parks.wrapper
}
