% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris fcns.R
\name{places.wrapper}
\alias{places.wrapper}
\title{places.wrapper}
\usage{
places.wrapper(countyfps = NULL, x = NULL, ...)
}
\arguments{
\item{countyfps}{countyfps to get parks for}

\item{x}{object to derive overlapping statefps from, if statefps is left as null. Aiddtional}
}
\description{
places.wrapper
}
