% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning-and-formatting.R
\name{quantile.bin_format}
\alias{quantile.bin_format}
\title{quantile.bin_format}
\usage{
\method{quantile}{bin_format}(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{
  Arguments passed on to \code{\link[=bin_from_breaks]{bin_from_breaks}}, \code{\link[=get.quantile.breaks]{get.quantile.breaks}}
  \describe{
    \item{\code{format_breaks}}{whether to create "x-y" type character labels for buckets rather
than default labels from \code{cut}}
    \item{\code{digits}}{Digits to pass on to \code{q.format} if using formatting}
    \item{\code{n_breaks}}{target number of breaks for x}
    \item{\code{highlight_top_percent}}{Whether to break out an additional bin for the top
1\%. Helpful to break out upper outliers.}
  }}
}
\description{
Bins a sequential variable based on value ranges. It can often make visualizations
easier to look at by allowing more striking differences and keeping outliers from
dominating the color scale. Wraps \code{get_breaks} and \code{get.quantile.breaks}, which can
also be used separately
}
