% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris fcns.R
\name{transform.and.crop}
\alias{transform.and.crop}
\title{transform.and.crop}
\usage{
\method{transform}{and.crop}(new.layer, basis.layer)
}
\arguments{
\item{basis.layer}{Layer or bbox to match with crs and crop}
}
\description{
Steps common to many of the wrappers if an sf object is supplied to one of the
tigris wrappers. Subsets and crops the input.
}
