% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times-series-plotting.R
\name{ts.labels.geom}
\alias{ts.labels.geom}
\title{ts.labels.geom}
\usage{
ts.labels.geom(tsx, group.col, date.col = "date", ...)
}
\arguments{
\item{tsx}{time series data}

\item{group.col}{column indicating group. Each will be labelled.}

\item{date.col}{date or time column for x axis.}

\item{...}{Passed onto \code{ggrepel::geom_label_repel}}
}
\description{
Returns a ggplot2 geometry that anticipates labeling lines of a multi-group
time series plot. Labels will appear at end of the plot.
}
