% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-themes-and-palettes.R
\name{upper.legend.box}
\alias{upper.legend.box}
\title{upper.legend.box}
\usage{
upper.legend.box(leg.pos = "right", lmargin = margin(4, 4, 4, 4))
}
\description{
Draws a box around a legend and top-justifies
}
