% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-device-helpers.R
\name{vector.ggsave}
\alias{vector.ggsave}
\title{vector.ggsave}
\usage{
vector.ggsave(
  plot,
  fn,
  sv.dir = "visuals/",
  ext = "svg",
  height = 7,
  width = height * 1.2,
  ...
)
}
\arguments{
\item{dir, fn}{directory and filename to save to}
}
\description{
Wraps ggsave with some defaults to save vectorized images.
}
