% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{water.wrapper}
\alias{water.wrapper}
\title{water.wrapper}
\usage{
water.wrapper(countyfps = NULL, x = NULL, size.min = 5e+06, ...)
}
\arguments{
\item{countyfps}{5-character state/county fp codes. Retrieved using
\code{county.subset} and supplied \code{x} argument if null.}

\item{x}{\code{sf} or \code{bbox} to get overlapping water areas for. Passed onto
\code{county.subset} if no county fps are supplied. Also used to spatially subset
water areas.}

\item{size.min}{Minimum size in m^2, after internal boundaries are resolved (if a
water area is represented by multiple contiguous polygons)}
}
\value{
water areas for region.
}
\description{
Gets water areas based on supplied countyfp codes and/or other spatial area.
}
