% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet helpers.R
\name{zoom_to_bounds}
\alias{zoom_to_bounds}
\title{zoom_to_bounds}
\usage{
zoom_to_bounds(x, mapID = "map")
}
\arguments{
\item{x}{sf object to zoom to on map. (Or one with equivalent bounds)}

\item{mapID}{ID for leaflet map}
}
\value{
Nothing, but references a leaflet proxy to update view bounds.
}
\description{
Flies to new boundary area
}
