% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping helpers.R
\name{add.map.layers}
\alias{add.map.layers}
\title{add.map.layers}
\usage{
add.map.layers(
  sfx,
  add.water = "#94bdff",
  add.counties = "#666666",
  add.places = "black",
  spatial.trim = c(st_intersection, st_crop),
  ...
)
}
\arguments{
\item{sfx}{sf object providing basis for map}

\item{add.water, add.counties}{add water areas/counties lines. Specify NULL if you
want to leave off this layer, otherwise specify a color.}

\item{spatial.trim}{Spatial trim method. Intersection or crop suggested.}

\item{p}{existing ggplot to add layers to. Blank ggplot is default}
}
\description{
Downloads map layers such as water and county lines and adds to a ggplot
}
