% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping helpers.R
\name{add.map.layers}
\alias{add.map.layers}
\title{add.map.layers}
\usage{
add.map.layers(sfx, p = ggplot(), add.water = T, add.counties = T)
}
\arguments{
\item{sfx}{sf object providing basis for map}

\item{p}{existing ggplot to add layers to. Blank ggplot is default}

\item{add.water, addcounties}{add water areas/counties lines}
}
\description{
Downloads map layers such as water and county lines and adds to a ggplot
}
