% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-helpers.R
\name{bbox2ggcrop}
\alias{bbox2ggcrop}
\title{bbox.to.ggcrop}
\usage{
bbox2ggcrop(sfx, crs = 4326)
}
\description{
Wraps \code{coord_sf}, which can be used to crop a map made in ggplot. Uses an sf object
or bbox to set crop area.
}
