% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-themes-and-palettes.R
\name{breaks_log10}
\alias{breaks_log10}
\title{breaks_log10}
\usage{
breaks_log10(x)
}
\description{
Manually adds breaks at very power of 10 for log-scale axes.
}
\details{
solution from https://r-graphics.org/recipe-axes-axis-log-ticks
}
