% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris fcns.R
\name{county.subset}
\alias{county.subset}
\title{county.subset}
\usage{
county.subset(x, cos = NULL, ...)
}
\arguments{
\item{x}{\code{sf} object}

\item{cos}{counties sf object. If none supplied, they are downloaded using
\code{tigris} library}

\item{...}{
  Arguments passed on to \code{\link[=flexible.spatial.filter]{flexible.spatial.filter}}
  \describe{
    \item{\code{subset.approach}}{spatial filter approach. Intersection or cropping to bbox.}
  }}
}
\description{
Given an \code{sf} object, get all counties that intersect. The spatial filter will
adapt based on input, but can be thrown off when \code{x} isn't coterminous with
counties. In this case the function \code{xwalks::get.spatial.overlap} can be helpful.
}
