% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris fcns.R
\name{county.subset}
\alias{county.subset}
\title{functions that interface with the \code{tigris} library to query layers.}
\usage{
county.subset(x, cos = NULL, spatial.filter = c("intersects", "crop"), ...)
}
\arguments{
\item{x}{\code{sf} object}

\item{cos}{counties sf object. If none supplied, they are downloaded using
\code{tigris} library}

\item{spatial.filter}{spatial filter approach. Intersection or cropping to bbox.}

\item{...}{passed onto \code{tigris::counties}}
}
\description{
The flow is: given a spatial layer, can query counties with county_subset, a thin
wrapper. Then other thin wrappers will call
county.subset
}
\details{
Given an \code{sf} object, get all counties that intersect. The spatial filter will
adapt based on input, but can be thrown off when \code{x} isn't coterminous with
counties. In this case the function \code{xwalks::get.spatial.overlap} can be helpful.
}
