% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-themes-and-palettes.R
\name{ggthme.logscales}
\alias{ggthme.logscales}
\title{ggthme.logscales}
\usage{
ggthme.logscales(axis.lbl.style = F)
}
\arguments{
\item{axis.lbl.style}{whether to apply \code{visaux::number.to.formatted.string}
to log axis labels}
}
\description{
ggthme.logscales
}
