% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-themes-and-palettes.R
\name{ggthme.logscales}
\alias{ggthme.logscales}
\title{ggthme.logscales}
\usage{
ggthme.logscales(
  break.intervals = list(x = c(1), y = c(1)),
  minor.breaks = list(x = minor.breaks_log10, y = minor.breaks_log10),
  axis.lbl.style = F
)
}
\arguments{
\item{break.intervals}{A vector of numerics between \link{1,9} for there to be
major breaks at every power of 10. I.e. every 1e6, 1e7 if left at just one,
but 1e6, 3e6, 1e6, 3e7 if includes \code{c(1, 3)}.}

\item{minor.breaks}{List with elements for x and y to pass onto
\code{minor_breaks} in the scale}

\item{axis.lbl.style}{whether to apply \code{visaux::number.to.formatted.string}
to log axis labels}
}
\description{
ggthme.logscales
}
