% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-device-helpers.R
\name{ragg.wrapper}
\alias{ragg.wrapper}
\title{ragg.wrapper}
\usage{
ragg.wrapper(fn, plot = ggplot2::last_plot(), sv.dir = "visuals/", ...)
}
\arguments{
\item{fn}{filename to save to}
}
\description{
Wraps \code{ragg::agg_png} to save a plot as .png. An alternative to ggsave. See
https://www.tidyverse.org/blog/2019/07/ragg-0-1-0/ for pkg intro.
}
